feat: delete comment

+++ src/application/usecases/comments/delete-comment.usecase.ts
@@ -1,18 +1,21 @@
+import { Role } from './../../../domain/model/vos/role.vo';
+import { User } from './../../../domain/model/entities/user.entity';
+import { IdVO } from './../../../domain/model/vos/id.vo';
+import { AuthorService } from './../../../domain/services/author.service';
 import { logger } from './../../../infrastructure/config/logger';
 import { ExceptionWithCode } from './../../../domain/model/exception-with-code';
 import { Post } from './../../../domain/model/entities/post.entity';
-import { IdVO } from '../../../domain/model/vos/id.vo';
 import { PostService } from './../../../domain/services/post.service';
 import { Service } from 'typedi';
 
 @Service()
 export class DeleteCommentUseCase {
 
-    constructor(private postService: PostService) {}
+    constructor(private postService: PostService, private authorService: AuthorService) {}
 
     async execute(request: DeleteCommentRequest): Promise<void> {
 
-        logger.debug(`Execute Delete Comment Use with ${JSON.stringify(request)}`);
+        logger.info(`Execute Delete Comment Use with ${JSON.stringify(request)}`);
 
         const post: Post | null = await this.postService.getById(IdVO.createWithUUID(request.idPost));
         if (!post) {
@@ -20,7 +23,15 @@ export class DeleteCommentUseCase {
         }
 
         const idComment = IdVO.createWithUUID(request.idComment);
-        
+        const comment = post.getCommentById(idComment);
+
+        const author = await this.authorService.getById(IdVO.createWithUUID(request.user.id.value));
+        logger.info(`${comment?.nickname.value} !== ${author?.nickname.value}`);
+
+        if (comment?.nickname.value !== author?.nickname.value && request.user.role.value !== Role.ADMIN) {
+            throw new ExceptionWithCode(403, 'Not allow delete this comment');
+        }
+
         await this.postService.deleteComment(post, idComment);
 
     }
@@ -30,4 +41,5 @@ export class DeleteCommentUseCase {
 export type DeleteCommentRequest = {
     idPost: string;
     idComment: string;
+    user: User;
 }
